requirements.txt
absl-py==1.0.0
astunparse==1.6.3
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cycler==0.11.0
deepface==0.0.68
easydict==1.9
filelock==3.4.0
Flask==2.0.2
flatbuffers==2.0
fonttools==4.28.3
gast==0.4.0
gdown==4.2.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
kornia==0.6.2
libclang==12.0.0
lpips==0.1.4
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
mtcnn==0.1.1
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
retina-face==0.0.5
rsa==4.8
scipy==1.7.3
six==1.16.0
soupsieve==2.3.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
typing-extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
